refactor: Start a design-system folder
