+++ src/components/YearlyContribution.js
@@ -2,13 +2,13 @@ import { useEffect, useLayoutEffect, useRef, useState, memo } from "react";
 import styled, { css } from "styled-components";
 import { space } from "@styled-system/space";
 
-import { Box } from "components/Box";
-import { Button } from "components/Button";
-import { Card, InfoCard } from "components/Card";
 import { DevIcon } from "components/DevIcon";
-import { Emoji } from "components/Emoji";
-import { Flex } from "components/Flex";
-import { Text } from "components/Text";
+import { Box } from "design-system/Box";
+import { Button } from "design-system/Button";
+import { Card, InfoCard } from "design-system/Card";
+import { Emoji } from "design-system/Emoji";
+import { Flex } from "design-system/Flex";
+import { Text } from "design-system/Text";
 
 import { useGitHub } from "hooks/useGitHub";
 import { useLastNonNullableValue } from "hooks/useLastNonNullableValue";
