+++ src/design-system/Emoji.tsx
